[refactor] #39 v2 버전 분리
