+++ src/main/java/com/abcdedu_backend/lecture/controller/LectureController.java
@@ -1,10 +1,6 @@
 package com.abcdedu_backend.lecture.controller;
 
 import com.abcdedu_backend.common.jwt.JwtValidation;
-import com.abcdedu_backend.lecture.dto.request.CreateAssignmentAnswerRequest;
-import com.abcdedu_backend.lecture.dto.request.CreateAssignmentRequest;
-import com.abcdedu_backend.lecture.dto.request.CreateLectureRequest;
-import com.abcdedu_backend.lecture.dto.request.CreateSubLectureRequest;
 import com.abcdedu_backend.lecture.dto.response.*;
 import com.abcdedu_backend.lecture.service.LectureService;
 import com.abcdedu_backend.utils.Response;
@@ -13,12 +9,8 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -40,63 +32,10 @@ public class LectureController {
 
     private final LectureService lectureService;
 
-    @ApiResponse(responseCode = "403", description = "api 권한이 없습니다. (admin만 가능)", content = @Content)
-    @Operation(summary = "클래스등록", description = "클래스(A, B, C, D)를 등록합니다.")
-    @PostMapping
-    public Response<Void> createLecture(@JwtValidation Long memberId,
-                                        @Valid @RequestBody CreateLectureRequest createLectureRequest){
-        lectureService.createLecture(memberId, createLectureRequest);
-        return Response.success();
-    }
-
     @Operation(summary = "클래스 조회", description = "클래스(A, B, C, D) 및 서브 클래스를 조회합니다.")
     @GetMapping
     public Response<List<GetClassResponse>> GetLectures(){
         List<GetClassResponse> response = lectureService.getLectures();
-        return Response.success(response);
-    }
-
-    @ApiResponse(responseCode = "403", description = "api 권한이 없습니다. (admin만 가능)", content = @Content)
-    @Operation(summary = "서브 클래스 등록", description = "서브 클래스를 등록합니다.")
-    @PostMapping("/{lectureId}")
-    public Response<Void> createSubLecture(@PathVariable Long lectureId,
-                                           @JwtValidation Long memberId,
-                                           @Valid @RequestBody CreateSubLectureRequest createSubLectureRequest){
-        lectureService.createSubLecture(lectureId, memberId, createSubLectureRequest);
-        return Response.success();
-    }
-
-    @ApiResponse(responseCode = "403", description = "api 권한이 없습니다. (admin만 가능)", content = @Content)
-    @Operation(summary = "시험 등록", description = "클래스 시험을 등록합니다.")
-    @PostMapping("/{subLectureId}/assignments")
-    public Response<Void> createAssignments(@PathVariable Long subLectureId,
-                                            @JwtValidation Long memberId,
-                                            @Valid @RequestBody CreateAssignmentRequest createAssignmentRequest){
-        lectureService.createAssignments(subLectureId, memberId, createAssignmentRequest);
-        return Response.success();
-    }
-
-    @Operation(summary = "시험 제출", description = "시험을 제출합니다.")
-    @PostMapping("/assignments/{assignmentId}")
-    public Response<Void> saveAssignmentAnswer(@PathVariable Long assignmentId,
-                                               @JwtValidation Long memberId,
-                                               @Valid @RequestBody CreateAssignmentAnswerRequest createAssignmentAnswerRequest){
-        lectureService.createAssignmentsAnswer(assignmentId, memberId, createAssignmentAnswerRequest);
-        return Response.success();
-    }
-
-    @Operation(summary = "시험 조회", description = "시험을 조회합니다.")
-    @GetMapping("/assignments/{assignmentId}")
-    public Response<GetAssignmentResponseV2> getAssignment(@PathVariable Long assignmentId){
-        GetAssignmentResponseV2 response = lectureService.getAssignment(assignmentId);
-        return Response.success(response);
-    }
-
-    @Operation(summary = "시험 제출 목록 조회 (admin)", description = "시험 제출 목록을 조회합니다.")
-    @GetMapping("/assignments/submissions")
-    public Response<List<GetAssignmentAnswerResponse>> getAssignmentAnswers(@PageableDefault(sort = {"createdAt"}, direction = Sort.Direction.DESC) Pageable pageable,
-                                                                            @JwtValidation Long memberId){
-        List<GetAssignmentAnswerResponse> response = lectureService.getAssignmentAnswers(pageable, memberId);
         return Response.success(response);
     }
 
@@ -128,6 +67,7 @@ public class LectureController {
         return Response.success(response);
     }
 
+    @ApiResponse(responseCode = "403", description = "api 권한이 없습니다.", content = @Content)
     @Operation(summary = "평가 파일 (시험/실습/프로젝트/이론) 조회", description = "평가 파일(시험/실습/프로젝트/이론)을 조회합니다.")
     @GetMapping("/file/{assignmentFileId}")
     public Response<GetAssignmentFileUrlResponse> getAssignmentFileUrl(@PathVariable Long assignmentFileId, @JwtValidation Long memberId){
@@ -135,6 +75,7 @@ public class LectureController {
         return Response.success(response);
     }
 
+    @ApiResponse(responseCode = "403", description = "api 권한이 없습니다.", content = @Content)
     @Operation(summary = "평가 파일 (시험/실습/프로젝트/이론) 문제지 조회", description = "평가 파일(시험/실습/프로젝트/이론) 문제지를 조회합니다.")
     @GetMapping("/answer-file/{assignmentAnswerFileId}")
     public Response<GetAssignmentAnswerFileUrlResponse> getAssignmentAnswerFileUrl(@PathVariable Long assignmentAnswerFileId, @JwtValidation Long memberId){
